Add tests for multipart builder
